+++ src/relstorage/adapters/packundo.py
@@ -23,14 +23,11 @@ from ZODB.POSException import UndoError
 from ZODB.utils import u64
 from zope.interface import implementer
 
-from .._compat import db_binary_to_bytes
 from ..iter import fetchmany
 from ..treemark import TreeMarker
 from .interfaces import IPackUndo
 
 # pylint:disable=too-many-lines,unused-argument
-
-
 
 
 log = logging.getLogger(__name__)
@@ -42,8 +39,8 @@ class PackUndo(object):
 
     _script_choose_pack_transaction = None
 
-    def __init__(self, database_type, connmanager, runner, locker, options):
-        self.database_type = database_type
+    def __init__(self, database_driver, connmanager, runner, locker, options):
+        self.driver = database_driver
         self.connmanager = connmanager
         self.runner = runner
         self.locker = locker
@@ -431,12 +428,9 @@ class HistoryPreservingPackUndo(PackUndo):
 
         add_rows = []  # [(from_oid, tid, to_oid)]
         for from_oid, state in self._fetchmany(cursor):
-            state = db_binary_to_bytes(state)
-            if hasattr(state, 'read'):
-                # Oracle
-                state = state.read() # pylint:disable=no-member
+            state = self.driver.binary_column_as_state_type(state)
             if state:
-                assert isinstance(state, bytes), type(state) # PY3: used to be str(state)
+                assert isinstance(state, self.driver.state_types), type(state)
                 from_count += 1
                 try:
                     to_oids = get_references(state)
@@ -956,13 +950,10 @@ class HistoryFreePackUndo(PackUndo):
         add_refs = []
 
         for from_oid, tid, state in self._fetchmany(cursor):
-            state = db_binary_to_bytes(state)
-            if hasattr(state, 'read'):
-                # Oracle
-                state = state.read() # pylint:disable=no-member
+            state = self.driver.binary_column_as_state_type(state)
             add_objects.append((from_oid, tid))
             if state:
-                assert isinstance(state, bytes), type(state)
+                assert isinstance(state, self.driver.state_types), type(state)
                 # XXX PY3 state = str(state)
                 try:
                     to_oids = get_references(state)