+++ src/relstorage/storage.py
@@ -602,11 +602,7 @@ class RelStorage(UndoLogCompatible,
                 state = self._adapter.mover.load_revision(
                     self._store_cursor, oid_int, tid_int)
 
-        if state is None:
-            raise POSKeyError(oid)
-
-        assert isinstance(state, bytes) # XXX PY3 used to do str(state)
-        if not state:
+        if state is None or not state:
             raise POSKeyError(oid)
         return state
 
@@ -648,7 +644,7 @@ class RelStorage(UndoLogCompatible,
                 end = p64(end_int)
             else:
                 end = None
-            assert isinstance(state, bytes), type(state) # XXX Py3 port: state = str(state)
+
             return state, p64(start_tid), end
 
     @Metric(method=True, rate=0.1)
@@ -1273,7 +1269,6 @@ class RelStorage(UndoLogCompatible,
             if not state:
                 return ()
 
-            assert isinstance(state, bytes), type(state) # XXX PY3: str(state)
             return {u64(oid) for oid in referencesf(state)}
 
         # Use a private connection (lock_conn and lock_cursor) to
@@ -1655,9 +1650,6 @@ class Record(DataRecord):
     """An object state in a transaction"""
 
     def __init__(self, tid, oid_int, data):
-        # XXX PY3: Used to to str(data) on Py2
-        if data is not None:
-            assert isinstance(data, bytes)
         DataRecord.__init__(self, p64(oid_int), tid, data, None)
 
 def _zlibstorage_new_instance(self):